# openapi.yaml
openapi: 3.0.3
info:
  title: Club Operations POS API
  version: 1.0.0
  description: |
    Server-authoritative REST API for managing club check-ins, room/locker assignments,
    cleaning workflows, checkout verification, and operational reporting.
    
    All endpoints are versioned under `/v1`.
    Request bodies are validated with Zod schemas.
    
    Authentication:
    - Staff endpoints require Bearer token (session token from PIN or WebAuthn login)
    - Admin endpoints require ADMIN role in addition to authentication
    - Customer-facing endpoints (agreement signing, checkout kiosk) do not require auth
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /v1/upgrades/waitlist:
    post:
      tags: [Upgrades]
      summary: Join waitlist for upgrade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId, desiredRoomType, acknowledgedDisclaimer]
              properties:
                sessionId:
                  type: string
                  format: uuid
                desiredRoomType:
                  type: string
                  enum: [STANDARD, DOUBLE, SPECIAL]
                acknowledgedDisclaimer:
                  type: boolean
      responses:
        '200':
          description: Waitlist entry created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session token from PIN or WebAuthn login
...